Update app.py
Event: On Huggingface
gradio.exceptions.DuplicateBlockError: A block with id: 1 has already been rendered in the current Blocks.

Action:
Modify textboxes

Observation: 
When you use the gr.Blocks() context manager, it creates a new scope for block IDs. When a new block is added, its ID is incremented from the last one. If you try to use a block with an ID that's already been used within the same gr.Blocks() scope, it will raise a DuplicateBlockError.